python-vision-service/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
opencv-python>=4.9.0
numpy>=1.26.0
pillow>=10.0.0
scipy>=1.11.0
requests>=2.31.0
